requirements.txt
# Core
Pillow==12.1.0
numpy==1.26.2
scipy==1.11.4

# GUI element detection (gui_locator.py)
opencv-python==4.10.0.84
pytesseract==0.3.13
